# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Private birthday time capsule API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UnlockInput {
  /** @minLength 1 */
  password: string;
}

export interface Session {
  authenticated: boolean;
}

export interface SiteConfig {
  personName: string;
  birthdayDate: string;
  introText: string;
}

export interface Memory {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  memoryDate: string | null;
  displayOrder: number;
}

export interface OpenWhenLetter {
  id: number;
  title: string;
  message: string;
  displayOrder: number;
}

export interface SiteContent {
  config: SiteConfig;
  memories: Memory[];
  openWhen: OpenWhenLetter[];
  finalLetter: string;
  birthdayMessage: string;
}

export interface Media {
  id: number;
  section: string;
  originalFilename: string;
  mimeType: string;
  size: number;
  objectPath: string;
}

export interface SnapshotAnswer {
  questionKey: string;
  questionText: string;
  answer: string;
  media?: Media[];
}

export interface Snapshot {
  id: number;
  isFinal: boolean;
  /** @nullable */
  completedAt: string | null;
  answers: SnapshotAnswer[];
}

export interface MediaInput {
  originalFilename: string;
  mimeType: string;
  size: number;
  objectPath: string;
  section: string;
}

export interface SnapshotAnswerInput {
  /** @minLength 1 */
  questionKey: string;
  /** @minLength 1 */
  questionText: string;
  answer: string;
  media?: MediaInput[];
}

export interface SnapshotInput {
  /** @minItems 1 */
  answers: SnapshotAnswerInput[];
}

export interface FutureLetter {
  id: number;
  letter: string;
  targetDate: string;
  isFinal: boolean;
}

export interface FutureLetterInput {
  /** @minLength 1 */
  letter: string;
}

export interface UploadInput {
  name: string;
  /** @minimum 1 */
  size: number;
  contentType: string;
}

export interface UploadResponse {
  uploadURL: string;
  objectPath: string;
}