# lib/api-zod/src/generated/types/futureLetterInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Private birthday time capsule API
 * OpenAPI spec version: 0.1.0
 */

export interface FutureLetterInput {
  /** @minLength 1 */
  letter: string;
}