# lib/api-zod/src/generated/types/media.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Private birthday time capsule API
 * OpenAPI spec version: 0.1.0
 */

export interface Media {
  id: number;
  section: string;
  originalFilename: string;
  mimeType: string;
  size: number;
  objectPath: string;
}