# lib/api-zod/src/generated/types/mediaInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Private birthday time capsule API
 * OpenAPI spec version: 0.1.0
 */

export interface MediaInput {
  originalFilename: string;
  mimeType: string;
  size: number;
  objectPath: string;
  section: string;
}