# lib/api-zod/src/generated/types/openWhenLetter.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Private birthday time capsule API
 * OpenAPI spec version: 0.1.0
 */

export interface OpenWhenLetter {
  id: number;
  title: string;
  message: string;
  displayOrder: number;
}