# lib/api-zod/src/generated/types/siteConfig.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Private birthday time capsule API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteConfig {
  personName: string;
  birthdayDate: Date;
  introText: string;
}