# lib/api-zod/src/generated/types/siteContent.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Private birthday time capsule API
 * OpenAPI spec version: 0.1.0
 */
import type { Memory } from './memory';
import type { OpenWhenLetter } from './openWhenLetter';
import type { SiteConfig } from './siteConfig';

export interface SiteContent {
  config: SiteConfig;
  memories: Memory[];
  openWhen: OpenWhenLetter[];
  finalLetter: string;
  birthdayMessage: string;
}