# lib/api-zod/src/generated/types/snapshot.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Private birthday time capsule API
 * OpenAPI spec version: 0.1.0
 */
import type { SnapshotAnswer } from './snapshotAnswer';

export interface Snapshot {
  id: number;
  isFinal: boolean;
  /** @nullable */
  completedAt: Date | null;
  answers: SnapshotAnswer[];
}