# lib/api-zod/src/generated/types/snapshotInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Private birthday time capsule API
 * OpenAPI spec version: 0.1.0
 */
import type { SnapshotAnswerInput } from './snapshotAnswerInput';

export interface SnapshotInput {
  /** @minItems 1 */
  answers: SnapshotAnswerInput[];
}