# lib/api-zod/src/generated/types/unlockInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Private birthday time capsule API
 * OpenAPI spec version: 0.1.0
 */

export interface UnlockInput {
  /** @minLength 1 */
  password: string;
}