# lib/api-zod/src/generated/types/uploadResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Private birthday time capsule API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadResponse {
  uploadURL: string;
  objectPath: string;
}